Move guard examples test to a separate file
